refactor. need to upload as is and switch to vm
need to upload as is. 
close to next release status. cli is not taking a list properly or as intended, need to figure that out, just a parsing syntax somewhere in the literals resolver. not ready to type safe out the urls with the regex function. when we do that is valid can be the type that we shove the string into after validation, similar to how its already being done.
on the gui side, things work but there is no interaction available for the multi threading, or passing in more than one song. I'm not sure I even want to hcnage that right now. gui workflow for most users would look heavy on the tube side,, ie making a playlist on youtube then passing the playlist in the downloader and downloading all at once. I do believe that will work currently if you pass in the youtu.be link for the playlist. pytube handles the links correctly, the question lies more in the regex validator I've bult infront of that.


also if not blatantly spelled out i guess that would be on me:
The program is now run form main, using an argument check to decide the mode that is spun up. None of this is testedw py installer. if anything the cli function from the the 0.1 release can stand as a utility right now and ill figure out a more versatile cli as this wraps up..
so gearing towards .2 with changes to the cli taking a list in arguments as somesort and adding all the features of the downloader, namely multi download to the gui if its simple enough. if not ill save that for the next iteration and keep the focus to the cli this pass.